+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @harmonicflow/docs
 
+## 2.0.4
+
+### Patch Changes
+
+- a
+- Updated dependencies
+  - @harmonicflow/react@3.0.3
+
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @harmonicflow/react@3.0.2
+
 ## 2.0.2
 
 ### Patch Changes