+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @harmonicflow/react
 
+## 3.0.3
+
+### Patch Changes
+
+- a
+
+## 3.0.2
+
+### Patch Changes
+
+- tet
+
 ## 3.0.1
 
 ### Patch Changes